requirements.txt
# API & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
python-multipart==0.0.6

# RAG & ML
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.35.2
faiss-cpu==1.7.4
huggingface-hub==0.16.4

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Utilities
pydantic==2.7.0
numpy<2,>=1.24.0
pandas<2,>=2.1.3
packaging<24,>=23.2
pillow<11,>=10.0.0
python-dotenv==1.0.0
requests==2.32.0

# Optional (API LLM)
openai==1.3.7
tiktoken==0.5.1
